# activity/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Activity Service API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/sessions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": [],
                        "GatewayKeyAuth": []
                    }
                ],
                "description": "Lists sessions with optional filters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sessions"
                ],
                "summary": "List sessions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by activity group ID",
                        "name": "activity_group_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by session template ID",
                        "name": "session_template_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status (scheduled, started, canceled, completed)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter by recurring status",
                        "name": "is_recurring",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by start time from (RFC3339)",
                        "name": "start_time_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by start time to (RFC3339)",
                        "name": "start_time_to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit number of results (default 20)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset for pagination (default 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sessions retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/api.Response-github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos_SessionListResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": [],
                        "GatewayKeyAuth": []
                    }
                ],
                "description": "Creates a new session for an activity group",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sessions"
                ],
                "summary": "Create a session",
                "parameters": [
                    {
                        "description": "Session data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos.CreateSessionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Session created successfully",
                        "schema": {
                            "$ref": "#/definitions/api.Response-github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos_CreateSessionResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            }
        },
        "/api/v1/sessions/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": [],
                        "GatewayKeyAuth": []
                    }
                ],
                "description": "Retrieves a session by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sessions"
                ],
                "summary": "Get a session",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Session retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/api.Response-github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos_SessionResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "404": {
                        "description": "Session not found",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": [],
                        "GatewayKeyAuth": []
                    }
                ],
                "description": "Updates an existing session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sessions"
                ],
                "summary": "Update a session",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos.UpdateSessionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Session updated successfully",
                        "schema": {
                            "$ref": "#/definitions/api.Response-github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos_EmptyResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "404": {
                        "description": "Session not found",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": [],
                        "GatewayKeyAuth": []
                    }
                ],
                "description": "Cancels a session (soft delete)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sessions"
                ],
                "summary": "Cancel a session",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Session cancelled successfully",
                        "schema": {
                            "$ref": "#/definitions/api.Response-github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos_EmptyResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "404": {
                        "description": "Session not found",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            }
        },
        "/api/v1/sessions/{id}/complete": {
            "post": {
                "security": [
                    {
                        "BearerAuth": [],
                        "GatewayKeyAuth": []
                    }
                ],
                "description": "Completes a session (changes status to completed)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sessions"
                ],
                "summary": "Complete a session",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Session completed successfully",
                        "schema": {
                            "$ref": "#/definitions/api.Response-github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos_EmptyResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "404": {
                        "description": "Session not found",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            }
        },
        "/api/v1/sessions/{id}/start": {
            "post": {
                "security": [
                    {
                        "BearerAuth": [],
                        "GatewayKeyAuth": []
                    }
                ],
                "description": "Starts a session (changes status to started)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sessions"
                ],
                "summary": "Start a session",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Session started successfully",
                        "schema": {
                            "$ref": "#/definitions/api.Response-github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos_EmptyResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "404": {
                        "description": "Session not found",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.Response-any": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "only on success"
                },
                "error": {
                    "description": "only on error",
                    "allOf": [
                        {
                            "$ref": "#/definitions/api.APIError"
                        }
                    ]
                },
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2020-01-01T00:00:00Z"
                }
            }
        },
...
        "api.Response-github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos_CreateSessionResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "only on success",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos.CreateSessionResponse"
                        }
                    ]
                },
                "error": {
                    "description": "only on error",
                    "allOf": [
                        {
                            "$ref": "#/definitions/api.APIError"
                        }
                    ]
                },
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2020-01-01T00:00:00Z"
                }
            }
        },
...
        "api.Response-github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos_SessionListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "only on success",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos.SessionListResponse"
                        }
                    ]
                },
                "error": {
                    "description": "only on error",
                    "allOf": [
                        {
                            "$ref": "#/definitions/api.APIError"
                        }
                    ]
                },
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2020-01-01T00:00:00Z"
                }
            }
        },
        "api.Response-github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos_SessionResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "only on success",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos.SessionResponse"
                        }
                    ]
                },
                "error": {
                    "description": "only on error",
                    "allOf": [
                        {
                            "$ref": "#/definitions/api.APIError"
                        }
                    ]
                },
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2020-01-01T00:00:00Z"
                }
            }
        },
...
        "github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos.CreateSessionRequest": {
            "type": "object",
            "required": [
                "activity_group_id",
                "location_city",
                "location_country",
                "location_lat",
                "location_lng",
                "start_time"
            ],
            "properties": {
                "activity_group_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426655440000"
                },
                "capacity": {
                    "type": "integer",
                    "maximum": 1000,
                    "minimum": 1,
                    "example": 20
                },
                "end_time": {
                    "type": "string",
                    "example": "2024-02-01T20:00:00Z"
                },
                "is_recurring": {
                    "type": "boolean",
                    "example": false
                },
                "location_city": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2,
                    "example": "Belgrade"
                },
                "location_country": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2,
                    "example": "Serbia"
                },
                "location_lat": {
                    "type": "number",
                    "maximum": 90,
                    "minimum": -90,
                    "example": 44.8176
                },
                "location_lng": {
                    "type": "number",
                    "maximum": 180,
                    "minimum": -180,
                    "example": 20.4633
                },
                "note": {
                    "type": "string",
                    "maxLength": 500,
                    "example": "Bring your own equipment"
                },
                "start_time": {
                    "type": "string",
                    "example": "2024-02-01T18:00:00Z"
                }
            }
        },
        "github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos.CreateSessionResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426655440000"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "GatewayKeyAuth": {
            "description": "Key added automatically by the gateway when proxying requests; required directly here only when bypassing the gateway.",
            "type": "apiKey",
            "name": "X-Gateway-Key",
            "in": "header"
        }
    },
    "security": [
        {
            "GatewayKeyAuth": []
        }
    ]
}